[Bronze V] Title: Can you add this?, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/7891. Can you add this？/Can you add this？.cc
@@ -9,7 +9,7 @@ int main() {
     int n;
     cin>>n;
     for(int  i=0;i<n;i++){
-        long long a,b;
+        int a,b;
         cin>>a>>b;
         cout<<a+b<<"\n";
     }
